+++ src/SGC4Lattice.h
@@ -1,5 +1,5 @@
-#ifndef SCT4Lattice_H
-#define SCT4Lattice_H
+#ifndef SGC4Lattice_H
+#define SGC4Lattice_H
 
 #include <vector>
 #include "CppLatticeUtils.h"
@@ -8,10 +8,10 @@
 #include <RcppThread.h>
 
 /**
- * @brief Computes directional spatial Granger causality between two spatial variables
+ * @brief Computes directional spatial granger causality between two spatial variables
  * on a spatial lattice using spatial neighbor embeddings and quantized entropy measures.
  *
- * This function quantifies the asymmetric spatial Granger causality between two
+ * This function quantifies the asymmetric spatial granger causality between two
  * spatial variables `x` and `y`, both defined over a spatial lattice structure.
  * It adopts an information-theoretic framework based on symbolic (or continuous)
  * entropy estimation, incorporating spatial embedding through neighboring structures.
@@ -55,11 +55,11 @@
  *
  * Returns:
  *   A `std::vector<double>` of size 2:
- *     - [0] Spatial Granger causality from `x` to `y`.
- *     - [1] Spatial Granger causality from `y` to `x`.
+ *     - [0] Spatial granger causality from `x` to `y`.
+ *     - [1] Spatial granger causality from `y` to `x`.
  *   If `normalize = true`, both values are scaled to the range [-1, 1].
  */
-std::vector<double> SCTSingle4Lattice(
+std::vector<double> SGCSingle4Lattice(
     const std::vector<double>& x,
     const std::vector<double>& y,
     const std::vector<std::vector<int>>& nb,
@@ -72,9 +72,9 @@ std::vector<double> SCTSingle4Lattice(
 );
 
 /**
- * @brief Compute spatial Granger causality for lattice data using spatial block bootstrap.
+ * @brief Compute spatial granger causality for lattice data using spatial block bootstrap.
  *
- * This function estimates the directional spatial Granger causality between two lattice variables `x` and `y`,
+ * This function estimates the directional spatial granger causality between two lattice variables `x` and `y`,
  * by applying a symbolic entropy-based method, and assesses the statistical significance of the causality using
  * spatial block bootstrap techniques. It calculates the causality in both directions: X → Y and Y → X.
  * Additionally, the function evaluates the significance of the estimated causality statistics by comparing them
@@ -111,12 +111,12 @@ std::vector<double> SCTSingle4Lattice(
  *
  * ## Returns:
  * A std::vector<double> of length 4:
- * - [0] Observed symbolic causality from x to y (sc_x_to_y).
+ * - [0] Observed spatial granger causality from x to y (sc_x_to_y).
  * - [1] Empirical p-value for x → y based on bootstrap distribution.
- * - [2] Observed symbolic causality from y to x (sc_y_to_x).
+ * - [2] Observed spatial granger causality from y to x (sc_y_to_x).
  * - [3] Empirical p-value for y → x based on bootstrap distribution.
  */
-std::vector<double> SCT4Lattice(
+std::vector<double> SGC4Lattice(
     const std::vector<double>& x,
     const std::vector<double>& y,
     const std::vector<std::vector<int>>& nb,
@@ -133,4 +133,4 @@ std::vector<double> SCT4Lattice(
     bool progressbar = true
 );
 
-#endif // SCT4Lattice_H
+#endif // SGC4Lattice_H